Picking up updates from Zillow's internal tycho.
Picking up an update to the events query that significantly improves
performance on tag and time-based queries. There are also some other minor
fixes outlined below.

The events used for unit tests had an expected ordering, which
no longer resolved correctly because the end_time was the same for all events,
causing the order by which they appear to be inconsistent.

Documented ordering expectations and modified unit tests to use the right order,
or not use them (there's a lot of unit tests testing the logic here).

There were a couple lingering tests based on schematics object instantiation,
and have been switched to attrs.