create_app is no longer a coroutine, nor is passed a loop
create_app being a coroutine can enable the mistake of instantiating
objects that need the loop on the construction of the app, rather than
as a startup call. This can cause issues, as the event loop should be
constructed after the worker is spawned to ensure no sharing of underlying
primitives such as kqueue or epoll.

create_app is not passed a loop anymore for similar reasons: instantion
of loop depending objects should occur during a startup signal.