Enabling configuration via environment variables (#2)
* Enabling configuration via environment variables

Exposing a way to configure the MongoDB connection URI via an
environment variable. This simplifies setup for the majority of
use cases, who simply would like to start a docker container and
pass in appropriate environment variables.

Internally tycho utilizes a config object, but exposing the existing
values would have severely limited the options available to configure
the MongoDB configuration.

Updating documentation to include instructions on how to bring up a
service against a non-local mongo.