Added retry on DuplicateKeyError for pymongo replace_one operation
There is a rare scenario where two threads may attempt the
replace/update operation at the same time.
Both threads will not find any match based on the query and
will try to insert the data.
One thread might update the data whereas
the other one will receive a DuplicateKeyError
because of the inserted data.
So to avoid this issue,
added a retry on occurance of DuplicateKeyError.
Adding a retry will return the previously inserted data
based on the query and help update the data.
Reference: https://jira.mongodb.org/browse/SERVER-14322