Use uranium-plus for build (#7)
using uranium-plus over orbital_core for
the build process. uranium-plus is a
best practice version of a ubuild file that
will stay compatible with the most recent version
of uranium, reducing our maintenance burden.

Newer versions of transmute no longer reject
keys that are non-existent in the schema, so
we need to refactor tests to use a valid key,
but an invalid value.